+++ docs/CHANGELOG.md
@@ -13,3 +13,4 @@ Changelog
 - Enh: Added filter to space calendar
 - Enh: Added calendar configuration
 - Fix: Added is public translation
+- Enh #17: Added ManageEntry permission for space members